// ssyweb/SSYNEW/src/app/models/pfInterestConfig.model.ts
export class PFInterestConfig {
    interestConfigId: number;
    interestCode: string;
    interestRate: any;
    effectiveDateFrom: Date;
    effectiveDateTo: Date;
    effectiveDateFromString: string;
    effectiveDateToString: string;
    status: number;
    cutoffDate: Date;
    description: string;
    interestCalculationDay: number;

}